macos_app/requirements.txt
PyQt6>=6.4.0
websocket-client>=1.3.3
requests>=2.28.1 